Revert build changes and test reorganization.
