+++ indexer/build.gradle
@@ -55,15 +55,14 @@ dependencies {
     testCompileOnly "com.google.code.findbugs:annotations:3.0.1u2"
 }
 
-sourceSets {
-    test {
-        resources {
-            srcDirs = [
-                "${rootDir}/underlay/src/testFixtures/resources",
-                "src/main/resources",
-                "src/test/resources"
-            ]
-        }
+test {
+    useJUnitPlatform()
+
+    // These System properties are used for the generated SQL strings to compare against during tests.
+    // See bio.terra.tanagra.aousynthetic.GeneratedSqlUtils for how they are used.
+    systemProperty("GRADLE_PROJECT_DIR", projectDir)
+    if (project.findProperty("generateSqlFiles")) {
+        systemProperty("GENERATE_SQL_FILES", "true")
     }
 }
 
@@ -74,7 +73,3 @@ task index (type:JavaExec) {
     systemProperties System.getProperties()
     args System.getProperty("exec.args", "").split()
 }
-
-// Apply Gradle Script Plugins for more modular gradle files.
-// See also https://docs.gradle.org/current/userguide/plugins.html#sec:script_plugins
-apply from: "$rootDir/underlay/gradle/task-dependencies.gradle"
