--- indexer/gradle/task-dependencies.gradle
@@ -1,15 +1,0 @@
-// Don't run any spotless or spotbugs tasks unless we have actual source file changes. The best
-// way to determine that is to rely on the compileJava determination.
-final TASKS_ONLY_IF_COMPILE_JAVA = [
-    spotlessApply,
-    spotlessJava,
-    spotlessJavaApply,
-    spotlessCheck
-]
-for (task in TASKS_ONLY_IF_COMPILE_JAVA) {
-    // Run every time the inputs to compilation change
-    task.dependsOn(compileJava.inputs.files)
-    // If compileJava runs, this must run after it (for the following to work).
-    task.mustRunAfter(compileJava)
-}
-
