+++ indexer/src/test/resources/config/entity/AttributeWithoutName.json
