+++ indexer/src/test/resources/config/entity/Brand.json
