+++ indexer/src/test/resources/config/entity/Condition.json
