+++ indexer/src/test/resources/config/entity/ConditionOccurrence.json
