+++ indexer/src/test/resources/config/entity/DoubleTextSearchMapping.json
