+++ indexer/src/test/resources/config/entity/EmptyAttributesListTextSearchMapping.json
