+++ indexer/src/test/resources/config/entity/EmptyTextSearchMapping.json
