+++ indexer/src/test/resources/config/entity/NoAttributes.json
