+++ indexer/src/test/resources/config/entity/NoIdAttribute.json
