+++ indexer/src/test/resources/config/entity/NoIndexDataMapping.json
