+++ indexer/src/test/resources/config/entity/NoSourceDataMapping.json
