+++ indexer/src/test/resources/config/entity/RawSqlTable.json
