+++ indexer/src/test/resources/config/entitygroup/BrandIngredient.json
