+++ indexer/src/test/resources/config/entitygroup/ConditionPersonOccurrence.json
