+++ indexer/src/test/resources/config/sql/condition_textsearch.sql
