+++ indexer/src/test/resources/generatedSql/condition_index_selectAllAttributes.sql
