+++ indexer/src/test/resources/generatedSql/condition_source_selectAllAttributes.sql
