+++ indexer/src/test/resources/generatedSql/person_index_selectAllAttributes.sql
