+++ indexer/src/test/resources/generatedSql/person_source_selectAllAttributes.sql
