+++ indexer/src/test/resources/generatedSql/rawsql_source_allInstances.sql
